[Refactor] #195 - MockNotificationData를 MockUserRepository에 통합
